### 2019-09-13 #1729 @whedon
[PRE REVIEW]: Pybotics: Python Toolbox for Robotics
**Submitting author:** @nnadeau (<a href="http://orcid.org/0000-0003-1220-0889">Nicholas A. Nadeau</a>)
**Repository:** <a href="https://github.com/nnadeau/pybotics" target ="_blank">https://github.com/nnadeau/pybotics</a>
**Version:** 0.7.1
