### 2019-09-13 #1729 @Kevin-Mattheus-Moerman
Perhaps reviewers from these:
https://github.com/openjournals/joss-reviews/issues/489
https://github.com/openjournals/joss-reviews/issues/500
https://github.com/openjournals/joss-reviews/issues/456
