### 2019-09-23 #1738 @adi3
@nnadeau Great work in the repo! I'm yet to test it out on my machine but I did a quick overview and I'm already a fan. Regarding the review, I'd like you to add a few things in the `Development` section of the readme.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Step 3 of your development process should have the contributor run tests to make sure the changes have not broken anything. Please list out specific commands and outputs the contributor should expect before the feature is ready for PR.

In addition, please add a quick line and a link to the issues board of your repo so users know where they can get support.
