+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,22 @@
+---
+name: Feature request
+about: Suggest an idea for this project
+title: "[FEATURE]"
+labels: enhancement
+assignees: ''
+
+---
+
+## User Story
+
+<!-- A clear and concise description of what the problem is. 
+E.g., As a User, I want to do X, because of Y. -->
+
+## Potential Solutions
+
+<!-- A clear and concise description of what you want to happen. -->
+<!-- A clear and concise description of any alternative solutions or features you've considered. -->
+
+## Additional Context
+
+<!-- Add any other context about the feature request here. -->
