### 2019-09-23 #1738 @nnadeau
@adi3 thanks for the review. The following changes have been made:

- Added issue templates: https://github.com/nnadeau/pybotics/commit/7d53e8f31756c316a0c84662a094a644bb376e3a
- As a [`CONTRIBUTING.md`](https://github.com/nnadeau/pybotics/blob/master/.github/CONTRIBUTING.md) already existed, I updated `README.md` with additional community info, direct links to the relevant docs/issues, and extended the testing how-to: https://github.com/nnadeau/pybotics/commit/2fe8b2127d560c676666c1a041d903716d3b18f8
