### 2019-09-23 #1738 @adi3
@nnadeau Thanks for updating that.

With reference to this quote from your paper:
> It was designed to provide a simple, clear, and concise interface to quickly simulate and evaluate common robot concepts, such as kinematics, dynamics, trajectory generations, and calibration.

I see the how the library helps with kinematics and calibration easily; both your examples and the source code show it clearly. I'm less certain about the `dynamics` and `trajectory generations` part of it. Of course, the kinematics and calibration features are important for calculating solutions of  dynamics and trajectory, but to actually mention it as an advantage of your library, please implement either (1) an out-of-the-box feature relating to them, or (2) an example script showing how the existing features lead to dynamics and planning equations.

Ditto for the mention of `Machine learning` in the Readme.
