### 2019-09-25 0eaf44c
added trajectory gen example
