### 2019-09-25 daade05
added machine learning example
