### 2019-09-25 9ea700b
added dynamics example
