### 2019-09-25 #1738 @nnadeau
@adi3 good catch! I've added additional examples (https://github.com/nnadeau/pybotics/pull/728). Specifically:

- https://github.com/nnadeau/pybotics/blob/master/examples/trajectory_generation.ipynb
- https://github.com/nnadeau/pybotics/blob/master/examples/machine_learning.ipynb
- https://github.com/nnadeau/pybotics/blob/master/examples/dynamics.ipynb

Example:

![image](https://user-images.githubusercontent.com/6395915/65566806-f04cf180-df21-11e9-9d1b-04b3eae55125.png)
