### 2019-09-25 #1738 @adi3
> @nnadeau Stellar work! I'm impressed by the utility of your library. I'll likely use it my work at some point.
>
> I'm having some trouble running your tests. This is the error I get: `No such file or directory: 'black': 'black' `. I'm not conversant with `poetry` so some pointers here would be useful.

My bad. Running `poetry install` got all this fixed. Your tests run fine and at this point, I am satisfied with your submission. Pending a references check, I will be happy to recommend your paper for publication.
