### 2019-09-25 #1738 @adi3
@nnadeau can you please update your .bib file to include the missing DOIs? See `whedon` output above.
