### 2019-09-25 9f8222e
added DOIs
