+++ paper/paper.bib
@@ -6,6 +6,7 @@
   number={2},
   pages={33},
   year={2019},
+  DOI={10.3390/robotics8020033},
   publisher={Multidisciplinary Digital Publishing Institute}
 }
 
@@ -14,6 +15,7 @@
   author={Corke, Peter},
   volume={118},
   year={2017},
+  DOI={10.1007/978-3-319-54413-7},
   publisher={Springer}
 }
 
@@ -33,6 +35,7 @@ year = {2005}
   number={2},
   pages={22},
   year={2011},
+  DOI={10.1109/MCSE.2011.37},
   publisher={IEEE Computer Society}
 }
 
@@ -49,6 +52,7 @@ year = {2005}
   booktitle={2018 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
   pages={1347--1353},
   year={2018},
+  DOI={10.1109/IROS.2018.8593598},
   organization={IEEE}
 }
 
