### 2019-09-25 #1738 @nnadeau
@adi3 @danielskatz DOIs updated: https://github.com/nnadeau/pybotics/pull/729

Missing DOIs:

- `Craig2005` (popular robotics textbook)
- `jones2001scipy` (recommended citation for library: https://www.scipy.org/citing.html#scipy-the-library)
- `pedregosa2011scikit` (recommended citation for library: https://scikit-learn.org/stable/about.html#citing-scikit-learn)
