### 2019-09-25 nnadeau/pybotics#730 @CameronDevine
Code of Conduct
There is a code of conduct in the repository, but it is not mentioned in the Readme or the Contributing instructions. I think this should be fixed to meet the requirements of openjournals/joss-reviews#1738.
