### 2019-09-25 nnadeau/pybotics#730 @nnadeau
Fixed in https://github.com/nnadeau/pybotics/commit/bcb62d974a14327af7eaaa48fe4033adc1fb52da
