### 2019-09-25 #1738 @nnadeau
@CameronDevine, https://github.com/nnadeau/pybotics/issues/730 is fixed. Thanks for the idea.
