### 2019-09-25 #1738 @CameronDevine
Once nnadeau/pybotics#731 is closed I will be ready to recommend this paper for publication.
