### 2019-09-26 8947d52
added docs requirements
