+++ docs/requirements.txt
@@ -1,0 +1,3 @@
+numpy
+scipy
+attrs
