+++ .readthedocs.yml
@@ -9,3 +9,4 @@ python:
   version: 3.7
   install:
     - requirements: docs/requirements.txt
+  
