### 2019-09-26 12d08e5
Update .readthedocs.yml
