+++ .readthedocs.yml
@@ -9,4 +9,3 @@ python:
   version: 3.7
   install:
     - requirements: docs/requirements.txt
-  
