### 2019-09-26 nnadeau/pybotics#731 @CameronDevine
In the link module there are a couple of documentation syntax errors where, `:return` is used instead of `:return:`. This causes return to be rendered incorrectly in the documentation.
