### 2019-09-26 #1738 @CameronDevine
@kyleniemeyer, I recommend this paper for publication.
