### 2019-09-28 #1738 @kyleniemeyer
Hi @nnadeau, just a few minor edits needed for the paper, in the references:

- the title in the Corke reference doesn't look completely correct, please check.
- The SciPy reference should be
```
@Misc{,
  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
  title =     {{SciPy}: Open source scientific tools for {Python}},
  year =      {2001--},
  url = "http://www.scipy.org/",
  note = {[Online; accessed <today>]}
}
```
- in the Nadeau 2018 reference, please put `{ }` around words that should be capitalized, like IEEE, RSJ, and IROS (or just the whole conference name)

We can ignore the warning about the missing DOI from whedon, that looks to be a review of the book you cited.
