+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @book{corke2017robotics,
-  title={Robotics, vision and control: fundamental algorithms in MATLAB{\textregistered} second, completely revised},
+  title={Robotics, Vision & Control},
   author={Corke, Peter},
   volume={118},
   year={2017},
