### 2019-09-28 609da08
using recommended citation for scipy
