+++ paper/paper.bib
@@ -39,11 +39,12 @@ year = {2005}
   publisher={IEEE Computer Society}
 }
 
-@article{jones2001scipy,
-  title={SciPy: Open source scientific tools for Python},
-  author={Jones, Eric and Oliphant, Travis and Peterson, Pearu and others},
-  year={2001},
-  publisher={https://www. scipy. org}
+@Misc{jones2001scipy,
+  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+  title =     {{SciPy}: Open source scientific tools for {Python}},
+  year =      {2001--},
+  url = "http://www.scipy.org/",
+  note = {[Online; accessed <today>]}
 }
 
 @inproceedings{nadeau2018evolutionary,
