### 2019-09-28 924dd59
added braces around conference title
