+++ paper/paper.bib
@@ -50,7 +50,7 @@ year = {2005}
 @inproceedings{nadeau2018evolutionary,
   title={Evolutionary motion control optimization in physical human-robot interaction},
   author={Nadeau, Nicholas A and Bonev, Ilian A},
-  booktitle={2018 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  booktitle={{2018 IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)}},
   pages={1347--1353},
   year={2018},
   DOI={10.1109/IROS.2018.8593598},
