+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @book{corke2017robotics,
-  title={{Robotics, Vision & Control}},
+  title={{Robotics, Vision \& Control}},
   author={Corke, Peter},
   volume={118},
   year={2017},
