### 2019-09-28 #1738 @nnadeau
@kyleniemeyer

- updated Corke citation title to match recommended title on [the toolbox's website](http://petercorke.com/wordpress/toolboxes/robotics-toolbox) (https://github.com/nnadeau/pybotics/commit/fe145a340f4d178c46a6abae523618756f4eda85)
- escaped ampersand in Corke title (https://github.com/nnadeau/pybotics/commit/58a4533baf4bbb4b9c59c78b000a3dc385245744)
- fixed scipy citation (https://github.com/nnadeau/pybotics/commit/609da085f9481c55fad66d7f63b233e2a2b221c0)
- added `{}` to Nadeau citation (https://github.com/nnadeau/pybotics/commit/924dd59c8dcc039ac860d646fbae7ec981a78e64)
