### 2019-09-28 #1738 @kyleniemeyer
@nnadeau thanks! The SciPy citation is still not quite right, since our template doesn't play well with the recommended year format. Perhaps just change it to "2001" ?
