+++ paper/paper.bib
@@ -42,7 +42,7 @@ year = {2005}
 @Misc{jones2001scipy,
   author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
   title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001--},
+  year =      {2001},
   url = "http://www.scipy.org/",
   note = {[Online; accessed <today>]}
 }
