### 2019-09-29 #1738 @kyleniemeyer
ah, whoops! @nnadeau, please archive your repo using a service such as Zenodo, and let me know what the associated DOI is.
