### 2019-09-30 a5b12f2
updated changelog for 0.7.2
