### 2019-09-30 #1738 @nnadeau
@kyleniemeyer

- Bumped release to [`0.7.2`](https://github.com/nnadeau/pybotics/releases/tag/0.7.2) (includes all changes from this review process)
- Fixed the the Zenodo integration (https://zenodo.org/record/3464672)
- Added DOI badge (https://github.com/nnadeau/pybotics/commit/43caa8de6bdb76e7aecc23e7d46e35e26f8f0834)
- DOI = `10.5281/zenodo.3464672 `
