### 2019-09-30 53e0e5e
set poetry version to git tag
