### 2019-09-30 #1738 @kyleniemeyer
@nnadeau sorry to drag this out, but it looks like you released v0.7.3—should that be what we use?
