### 2019-09-30 #1738 @nnadeau
@kyleniemeyer sure! that was outside the scope of this review process (i.e., travis CI update), but sure, go ahead!

DOI:
    10.5281/zenodo.3464685
