### 2019-09-30 #1738 @kyleniemeyer
Congrats @nnadeau on your article's publication in JOSS! Thanks to @adi3 and @CameronDevine for reviewing!
